requirements-serverless.txt
pdf2image==1.16.3
paddleocr==2.7.0.3
paddlepaddle-gpu==2.6.2  # GPU-enabled version (CUDA 11.7)
opencv-python-headless==4.6.0.66  # Headless version for serverless
numpy<2.0
runpod>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
